clean up. tests passing now
